refactor(Controllers): Huge controllers refactoring
Moved movie editing to another new controller.
Improved exceptions
checking.
Attribute for requiring model to be valid.

+++ webapi/Controllers/ReviewController.cs
@@ -1,39 +1,44 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using MovieCatalogBackend.Data.MovieCatalog;
 using MovieCatalogBackend.Data.MovieCatalog.Dtos;
+using MovieCatalogBackend.Exceptions;
 using MovieCatalogBackend.Helpers;
-using Oracle.ManagedDataAccess.Client;
 
 namespace MovieCatalogBackend.Controllers;
 
 [ApiController]
+[RequireValidModel]
 [Route("api/movie/{movieId:guid}/review")]
 public class ReviewController : ControllerBase
 {
     private readonly MovieCatalogContext _context;
     private readonly ILogger _logger;
+    private readonly IDbExceptionsHelper _exHelper;
 
-    public ReviewController(MovieCatalogContext context, ILogger<ReviewController> logger)
+    public ReviewController(MovieCatalogContext context, IDbExceptionsHelper exHelper, ILogger<ReviewController> logger)
     {
         _context = context;
         _logger = logger;
+        _exHelper = exHelper;
     }
 
     [Authorize(Policy = "TokenNotBlacked")]
+    [Authorize(Policy = "ReviewerPermissions")]
     [HttpPost("add")]
     public async Task<ActionResult> Add(Guid movieId, ReviewModifyModel model)
     {
-        if (!User.SidAsGuid(out var userId)) return Unauthorized();
+        if (!User.TryGetSidAsGuid(out var userId)) return Unauthorized();
         try
         {
             _context.Review.Add(model.ToReview(userId, movieId));
             await _context.SaveChangesAsync();
         }
-        catch (DbUpdateException e) when (e.InnerException is OracleException { Number: 1 })
+        catch (Exception e) when (_exHelper.IsAlreadyExist(e))
         {
-            return Problem("User has already had review for this movie", statusCode: StatusCodes.Status409Conflict);
+            return Problem(title: _exHelper.Message, detail: "User has already had review for this movie",
+                statusCode: StatusCodes.Status409Conflict);
         }
         catch (Exception e)
         {
@@ -48,20 +53,21 @@ public class ReviewController : ControllerBase
     [HttpDelete("{reviewId}/delete")]
     public async Task<ActionResult> Delete(Guid movieId, Guid reviewId)
     {
+        if (!User.TryGetSidAsGuid(out var userId)) return Unauthorized();
         try
         {
-            _context.Review.Remove(new Review { Id = reviewId }); 
+            _context.Review.Remove(new Review { Id = reviewId });
             await _context.SaveChangesAsync();
             return Ok();
         }
-        catch (DbUpdateConcurrencyException e)
+        catch (Exception e) when (_exHelper.IsNotFound(e))
         {
             _logger.LogWarning(e, $"Occured while deleting review({reviewId})");
-            return Problem(title: "Unknown review id", statusCode: StatusCodes.Status404NotFound);
+            return Problem(title: _exHelper.Message, detail: "Unknown review id",
+                statusCode: StatusCodes.Status404NotFound);
         }
         catch (Exception e)
         {
-            User.SidAsGuid(out var userId);
             _logger.LogError(e, $"Unknown exception while deleting review from user({userId}) on movie({movieId})");
             return Problem(title: "Unexpected exception occured");
         }
@@ -71,17 +77,18 @@ public class ReviewController : ControllerBase
     [HttpPost("{reviewId}/edit")]
     public async Task<ActionResult> Edit(Guid movieId, Guid reviewId, ReviewModifyModel model)
     {
-        if (!User.SidAsGuid(out var userId)) return Unauthorized();
+        if (!User.TryGetSidAsGuid(out var userId)) return Unauthorized();
         try
         {
             _context.Review.Update(model.ToReview(userId, movieId, reviewId));
             await _context.SaveChangesAsync();
             return Ok();
         }
-        catch (DbUpdateConcurrencyException e)
+        catch (Exception e) when (_exHelper.IsNotFound(e))
         {
             _logger.LogWarning(e, $"Occured while editing review({reviewId})");
-            return Problem(title: "Unknown review id", statusCode: StatusCodes.Status404NotFound);
+            return Problem(title: _exHelper.Message, detail: "Unknown review id",
+                statusCode: StatusCodes.Status404NotFound);
         }
         catch (Exception e)
         {
